Attempt to address most of Pete's comments on the pull request
