+++ profile_bluesky/startup/instrument/devices/__init__.py
@@ -10,9 +10,9 @@ from .monochromator import *
 from .huber import *
 from .slits import *
 from .phaseplates import *
-from .others import *
 from .lakeshores import *
-
-# under development
-# from .TODO.fourc import *
-# from .TODO.kappa_config import *
+from flat_mirror import *
+from toroidal_mirror import *
+from xbpm import *
+from uptable import *
+from kbic import *
