+++ profile_bluesky/startup/instrument/devices/magnets.py
@@ -15,22 +15,22 @@ from ophyd import Component, Device, EpicsMotor, FormattedComponent
 class Magnet6T(Device):
 
     ## Motors ##
-    tabth = Component(EpicsMotor,'m53', labels=('motor','6T magnet'))  # 4T Mag Th
-    tabx = Component(EpicsMotor,'m49', labels=('motor','6T magnet'))  # 4T MagTab X
-    taby = Component(EpicsMotor,'m50', labels=('motor','6T magnet'))  # 4T MagTab Y
+    tabth = Component(EpicsMotor,'m53', labels=('motor','6T_magnet'))  # 4T Mag Th
+    tabx = Component(EpicsMotor,'m49', labels=('motor','6T_magnet'))  # 4T MagTab X
+    taby = Component(EpicsMotor,'m50', labels=('motor','6T_magnet'))  # 4T MagTab Y
 
-    tabth2 = Component(EpicsMotor,'m56', labels=('motor','6T magnet'))  # AMIMagnetPhi
-    tabz2 = Component(EpicsMotor,'m51', labels=('motor','6T magnet'))  # AMIMagnetZ
-    tabx2 = Component(EpicsMotor,'m52', labels=('motor','6T magnet'))  # AMIMagenetX
+    tabth2 = Component(EpicsMotor,'m56', labels=('motor','6T_magnet'))  # AMIMagnetPhi
+    tabz2 = Component(EpicsMotor,'m51', labels=('motor','6T_magnet'))  # AMIMagnetZ
+    tabx2 = Component(EpicsMotor,'m52', labels=('motor','6T_magnet'))  # AMIMagenetX
 
-    sampy = Component(EpicsMotor,'m63', labels=('motor','6T magnet'))  # CIATRA
-    sampth = Component(EpicsMotor,'m58', labels=('motor','6T magnet'))  # CIA ROT
+    sampy = Component(EpicsMotor,'m63', labels=('motor','6T_magnet'))  # CIATRA
+    sampth = Component(EpicsMotor,'m58', labels=('motor','6T_magnet'))  # CIA ROT
 
     Tvaporizer = None
     Tsample = None
 
 
-mag6t = Magnet6T('4iddx:',name='6T magnet')
+mag6t = Magnet6T('4iddx:',name='6T_magnet')
 # Tvaporizer = lakeshore_336.loop1
 # Tsample = lakeshore_336.loop2
